nightlydevzip: option to build test branch

+++ nightlydevzip/setup.sh
@@ -6,12 +6,16 @@ rpm --import "http://keyserver.ubuntu.com/pks/lookup?op=get&search=0x3FA7E032808
 rpm --import "http://keyserver.ubuntu.com/pks/lookup?op=get&fingerprint=on&search=0x4796B710919684AC"
 yum -y install mono-devel libgdiplus-devel xsp mono-mvc mono-data-sqlite liberation-mono-fonts nant wget tar sqlite php-cli curl gettext libsodium
 
-wget https://github.com/openpetra/openpetra/archive/master.tar.gz || exit -1
-tar xzf master.tar.gz
-mv openpetra-master nightlydevzip
-#wget https://github.com/tpokorra/openpetra/archive/mono4.tar.gz || exit -1
-#tar xzf mono4.tar.gz
-#mv openpetra-mono4 nightlydevzip
+if [[ "$branch" == "master" ]]
+then
+  wget https://github.com/openpetra/openpetra/archive/$branch.tar.gz -O sources.tar.gz || exit -1
+else
+  wget https://github.com/tpokorra/openpetra/archive/$branch.tar.gz -O sources.tar.gz || exit -1
+fi
+
+tar xzf sources.tar.gz
+dir=$(find . -type d -name openpetra-*)
+mv $dir nightlydevzip
 
 cd nightlydevzip
 cat > OpenPetra.build.config << EOF
@@ -27,13 +31,18 @@ rm -Rf csharp/ThirdParty/SQLite/Mono.Data.Sqlite.dll
 cp /usr/lib/mono/4.5/Mono.Data.Sqlite.dll csharp/ThirdParty/SQLite/
 
 nant devzip || exit -1
-mv ../openpetra_development_`date +"%Y-%m-%d"`.zip ~/tarball
 
-#upload to Github
-if [ -f ~/.ssh/github_config.php ]
+if [[ "$branch" == "master" ]]
 then
-  php ../github_uploadrelease.php || exit -1
-fi
+  # only upload builds of master
+  mv ../openpetra_development_`date +"%Y-%m-%d"`.zip ~/tarball
 
-rm -f ~/tarball/openpetra_development_`date +"%Y-%m-%d" --date='6 days ago'`.zip
-echo download at https://download.solidcharity.com/tarballs/tpokorra/openpetra/openpetra_development_`date +"%Y-%m-%d"`.zip
+  #upload to Github
+  if [ -f ~/.ssh/github_config.php ]
+  then
+    php ../github_uploadrelease.php || exit -1
+  fi
+
+  rm -f ~/tarball/openpetra_development_`date +"%Y-%m-%d" --date='6 days ago'`.zip
+  echo download at https://download.solidcharity.com/tarballs/tpokorra/openpetra/openpetra_development_`date +"%Y-%m-%d"`.zip
+fi
